Prevent CCMs to be on the same node
If CCMs are deployed on the same node, only one can start successfully.
Other are in CrashLoopBackOff with this error message:
error: failed to create listener: failed to listen on 127.0.0.1:10258: listen tcp 127.0.0.1:10258: bind: address already in use

To prevent them to be scheduled together this commit sets a stricter
podAntiAffinity type: requiredDuringSchedulingIgnoredDuringExecution.