Use just one CCM replica for Single Node deployments
In case of Single Node deployments we need to create just one CCM
replica, because there will be a port conflict if there are many
of them.

To do so we observe the Infrastructure resource to check what kind
of control plane topology the cluster has. If it's SingleReplicaTopologyMode
then the operator creates just one CCM replica.