[OCPCLOUD-1202] Allow CCMs to observe cluster wide proxy
To conform with existing components within OpenShift, this commit
allows the operator to watch for ClusterWideProxy configuration and
inject the relevant environment variables into the containers for
the CCMs.

This ensures that cloud provider traffic is routed via customers
proxies as per the rest of the outbound traffic from the OpenShift
components.