Fix fmt test
Currently we expect that "make fmt" just checks the code and returns
an error when it's badly formatted. Unfortunately, it just fixes the
code and finishes successfully.

To fix if, this commit introduces a new script "verify-gofmt.sh" that
is used in the original "fmt" target. The "fmt" itself was renamed
to "update-fmt" to comply with what it does.