Check diff after generate and vendor tasks
Now our targets just fix related issues in the code. For CI it means
that it doesn't report any errors, as jobs always finish successfully.

This commit adds a script that checks the output of git diff command.
If the diff is not empty, then the code has been modified, and therefore
it has some errors.

Running this script in CI will force it to report real errors in the
jobs logs.

Co-authored-by: Danil Grigorev <dgrigore@redhat.com>