Add tests to comply with OpenShift requirements
These tests allow to make sure that the operator applies the resources
as required by OpenShift: users are allowed to add custom labels and
annotations, but everything else will be overwritten by the operator.

Since we use Server-Side Apply mechanism that sometimes doesn't comply
with these requirements, it should be replaced to prevent collisions.