controllers: Add logic for transforming CCM config
We're attempting to switch OpenStack clouds (and eventually, other
clouds) from the legacy cloud provider to external cloud providers
called by the Cloud Controller Manager (CCM) [1]. The cloud config
operator (CCO) is currently responsible for managing the configuration
of the legacy cloud provider. To do this, it implements transformation
functionality which allows it to transform user-provided configuration
to operator-managed configuration. For many clouds, this transformation
simply involves ensuring the configuration exists under a standard
(rather than cloud-specific) key in the resulting configmap [2],
however, for other clouds (currently AWS and Azure) the transformation
is a little more involved [3][4].

As we transition to external cloud providers, CCCMO is going to take
over the responsibility of managing configuration of the newly external
cloud provider service. This means CCCMO is going to have to support
transforming the user-supplied configuration like CCO does. On top of
the existing transformation steps though, it's also going to need to
account for the fact that configuration for the external cloud providers
may (and certainly does, in the OpenStack case) differ between legacy
and external cloud providers. This means additional transformation and a
level of validation will be necessary to allow migration to CCM without
forcing the user to manually fix their configurations.

In this change, we add the infrastructure that will eventually allow us
to implement this transformation logic on a per-cloud basic. Currently
the only transformer implemented is the 'NoOpTransformer' which simply
ensures configuration lives under the standard key. However, future
changes will add cloud-specific transformers to both cover what CCO was
previously doing and handle the differences between the legacy cloud
provider configuration and external cloud provider configuration.

[1] https://kubernetes.io/docs/concepts/architecture/cloud-controller/
[2] https://github.com/openshift/cluster-config-operator/blob/19ff048/pkg/operator/kube_cloud_config/controller.go#L137-L153
[3] https://github.com/openshift/cluster-config-operator/blob/19ff048/pkg/operator/kube_cloud_config/aws.go#L16-L84
[4] https://github.com/openshift/cluster-config-operator/blob/19ff048/pkg/operator/kube_cloud_config/azure.go#L40-L120

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>