cloud: Implement OpenStack transformer logic
Add support for transforming user-provided configuration. As discussed
previously, there are two tasks required here: ensuring the
configuration lives under a well-known key and ensuring that any
configuration options for the legacy cloud provider are removed and
replaced by configuration options for the external cloud provider. As
discussed in the enhancement [1], the key difference is that the
external OpenStack cloud provider relies on a local 'clouds.yaml' file
rather than a secret containing the contents of such a file. It also
removes some legacy configuration options. In summary, we need to drop
the following configuration options (after ensuring they haven't been
modified from the defaults):

- [Global] secret-namespace
- [Global] secret-name
- [Global] kubeconfig-path

...and add the following:

- [Global] use-clouds
- [Global] clouds-file
- [Global] cloud

We do just that.

[1] https://github.com/openshift/enhancements/pull/1009

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>