update enabled and use-octavia options in cloud-conf config-map
Since Neutron-LBaaS has been deprecated in OpenStack queens release,
we must ensure installations use Octavia by default. Also, when running
installations with Kuryr SDN the cloud provider shouldn't handle the
Services as Kuryr already does that with Octavia. This commit fixes the
issues by ensuring that the configurations use-octavia is set as True and
enabled is set as False, when the cluster network is configured with Kuryr.