OpenStack: ensure config-map is updated only when needed
The managed config-map is often updated even when all the
keys have the correct value, this behaviour is caused
by the fact that the order of the iteration of config-map
keys not preserved when using golang Map, consequently the final
version of the config-map would be considered different
from the previous one, when shouldn't. This commit fixes
the issue by ensuring that the order of the keys is
preserved in the config-map.