vSphere: If one failuredomain do not config labels
Changes to the installer to GA zonal
always create at least one failuredomain.
If labels are set and the tag or tag
categories do not exist installation boostrap
fails. To support the legacy non-zonal install
labels will only be applied when the length of
failure domains is greater than one.