add a check for nutanix cloud conf map
When upgrading from 4.12 to 4.13 on the nutanix platform, the cloud-conf
configmap must be present to avoid a degradation. This change adds a
check in the cloud_config_sync_controller reconcile to check the
platform type and configmap existence, if it does not exist on nutanix
then the upgradeable condition will be set to false.

ref: https://issues.redhat.com/browse/OCPBUGS-7898