add separate upgradeable condition for the sync controller
this change adds a separate condition for the sync controller so that
primary controller can more accurately detect when the sync controller
wants to mark a blocked upgrade. Separating the upgradeable condition in
this manner prevents the primary controller from becoming blocked if it
goes non-upgradeable due to its own issues or degraded status.

added a unit test to check that the interaction between the config sync
controller and the primary controller is working the way we want.