update return values from GetCloudConfigTransformer
This change adds a boolean return value to the GetCloudConfigTransformer
function to indicate when a the cloud config ConfigMap should be looked
up from the CCO namespace. It is being added so that we can return a
config transformer and also be able to lookup the default values from
the CCO. This is primarily to address OCPBUGS-20213 which requires an
updated config on Azure Stack Hub.