add a cloud config transformer for azure stack
This change creates a CloudConfigTransformer for Azure Stack cloud
provider. The transformer will unmarshal json from the cloud config into
a `Config` struct from the azure cloud provider and then update the the
`VMType` field if it is unset. This is being added because Azure has
updated the default vm type to value that cannot be added to the basic
load balancers which are associated with Azure Stack.

The change also adds a test suite to accompany the transformer.