Adds CloudConfigTransformer for Azure
This adds a CloudConfigTransformer for Azure that will set the VMType to
'standard' if unset.

See #291, OCPBUGS-25483 and OCPBUGS-20213 for more information.

This change also adds the functionality from the cluster-config-operator
azure CloudConfigTransformer.

The functionality we bring over is:

1. Ensure that the Cloud is set in the cloud.conf
   i. If it is set, verify that it is valid and does not conflict with the
      infrastructure config. If it conflicts, we want to error
  ii. If it is not set, default to public cloud (configv1.AzurePublicCloud)

 2. Verify the cloud name set in the infra config is valid, if it is not
 bail with an informative error

See openshift/cluster-config-operator#133 for
the PR that initially added this functionality.