Pass --cluster-name to OpenStack CCM
By an terrible oversight, since the in-tree, OpenStack cloud provider
was run without passing `--cluster-name` argument, meaning that it just
used the default of "kubernetes" as the cluster name when constructing
the names of the LB resources. Obviously this led to conflicts when
there were multiple OCP clusters in a single OpenStack tenant and it's
still a challenge for example for QE running tests on a single cloud.

This issue couldn't be fixed easily, because we haven't had an upgrade
story solved. Now kubernetes/cloud-provider-openstack#2552 is an attempt
to make CCM rename the LBs when the cluster-name is reconfigured. This
should allow us to finally solve the name conflict issues.

This commit makes sure that CCMO deploys OpenStack CCM passing
`--cluster-name`.